+++ db/migrate/20250627000002_add_note_to_thing_links.rb
@@ -1,0 +1,5 @@
+class AddNoteToThingLinks < ActiveRecord::Migration[8.1]
+  def change
+    add_column :thing_links, :note, :text
+  end
+end