+++ db/migrate/20250627000003_add_ar_anchor_note_to_things.rb
@@ -1,0 +1,5 @@
+class AddArAnchorNoteToThings < ActiveRecord::Migration[8.1]
+  def change
+    add_column :things, :ar_anchor_note, :text
+  end
+end