+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2025_06_27_000001) do
+ActiveRecord::Schema[8.1].define(version: 2025_06_27_000003) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -70,6 +70,7 @@ ActiveRecord::Schema[8.1].define(version: 2025_06_27_000001) do
   create_table "thing_links", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.string "link_type", null: false
+    t.text "note"
     t.integer "position"
     t.bigint "thing_id", null: false
     t.string "title"
@@ -81,6 +82,7 @@ ActiveRecord::Schema[8.1].define(version: 2025_06_27_000001) do
   end
 
   create_table "things", force: :cascade do |t|
+    t.text "ar_anchor_note"
     t.datetime "created_at", null: false
     t.text "description"
     t.string "ip_address"